🔐 Added safety measures on forms

+++ includes/delete-account.inc.php
@@ -1,14 +1,20 @@
 <?php
+include_once('./user-role-check.inc.php');
 
-if($_SERVER["REQUEST_METHOD"] == "POST")
-{    
-    sleep(2);
-
-    if (isset($_POST['user_id'])) {
-        $userId = $_POST['user_id'];
-        $userManager->deleteUser($userId);
-    } else {
-        header("Location: ../access-admin-logma/delete-account?error=missinguserid");
-        exit();
+if($userAdmin){
+    if($_SERVER["REQUEST_METHOD"] == "POST")
+    {    
+        sleep(2);
+    
+        if (isset($_POST['user_id'])) {
+            $userId = $_POST['user_id'];
+            $userManager->deleteUser($userId);
+        } else {
+            header("Location: ../access-admin-logma/delete-account?error=stmtfailed");
+            exit();
+        }
     }
-}
+} else{
+    $sessionManager->forbiddenAccess();
+} 
+
